[Utils] Big restructuring
