+++ hottbox/datasets/utils/tests/test_special.py
@@ -2,9 +2,9 @@ import pytest
 import numpy as np
 from ..special import *
 from ....core.structures import Tensor
-from ....utils import isToepTensor, isToepMatrix
+from ....utils.checks import isToepTensor, isToepMatrix
 
-def test_toeplitzTensor():
+def test_toeplitz():
     tensor = np.zeros(shape=(4,4,3))
     # Inititalise
     mat_A = genToeplitzMatrix([1,2,3,4],[1,4,3,2])
@@ -14,12 +14,12 @@ def test_toeplitzTensor():
     tensor[:,:,1] = mat_B
     tensor[:,:,2] = mat_C
 
-    tt = toeplitzTensor((4,4,3), matC=np.array([mat_A, mat_B, mat_C])).data
+    tt = toeplitz((4,4,3), matC=np.array([mat_A, mat_B, mat_C])).data
     assert np.array_equal(tt, tensor)
 
 
-def test_toeplitzTensorRandom():
-    test_tensor = toeplitzTensor((3,3,4), modes=[0,1], random=True)
+def test_toeplitz_random():
+    test_tensor = toeplitz((3,3,4), modes=[0,1], random=True)
     assert isToepTensor(test_tensor, modes=[0,1])
