+++ hottbox/utils/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Module for generation of synthetic N-dimensional arrays (tensors)
+"""
